Разделила код на отдельные модули
